add readme and change to easyocr
