+++ pdf/easyocr/requirements.txt
