+++ pdf/easyocr/utils/ocr_module.py
